missed some short -> int conversion
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@1058 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/security/CERTConverter.java
@@ -49,7 +49,7 @@ parseRecord(CERTRecord r) {
 
 /** Builds a CERT record from a Certificate associated with a key also in DNS */
 public static CERTRecord
-buildRecord(Name name, short dclass, int ttl, Certificate cert, int tag,
+buildRecord(Name name, int dclass, int ttl, Certificate cert, int tag,
 	    int alg)
 {
 	short type;
@@ -74,7 +74,7 @@ buildRecord(Name name, short dclass, int ttl, Certificate cert, int tag,
 
 /** Builds a CERT record from a Certificate */
 public static CERTRecord
-buildRecord(Name name, short dclass, int ttl, Certificate cert) {
+buildRecord(Name name, int dclass, int ttl, Certificate cert) {
 	return buildRecord(name, dclass, ttl, cert, 0, 0);
 }
